Refactored the creation of FractionalRebinning Quad
Since finding that the quadrilateral needed for Fractional Rebinning
is not consistent for each type of conversion, it has been decided that
this code should be refactored into the CalculateReflectometry header for
each specific conversion.

Refs #13411
