+++ Code/Mantid/Framework/DataObjects/CMakeLists.txt
@@ -63,7 +63,9 @@ set ( INC_FILES
 	inc/MantidDataObjects/AffineMatrixParameterParser.h
 	inc/MantidDataObjects/BoxControllerNeXusIO.h
 	inc/MantidDataObjects/CalculateReflectometry.h
+    inc/MantidDataObjects/CalculateReflectometryKiKf.h
 	inc/MantidDataObjects/CalculateReflectometryQxQz.h
+    inc/MantidDataObjects/CalculateReflectometryP.h
 	inc/MantidDataObjects/CoordTransformAffine.h
 	inc/MantidDataObjects/CoordTransformAffineParser.h
 	inc/MantidDataObjects/CoordTransformAligned.h
