+++ Code/Mantid/Framework/MDAlgorithms/inc/MantidMDAlgorithms/ReflectometryTransformKiKf.h
@@ -1,75 +1,16 @@
 #ifndef MANTID_MDALGORITHMS_REFLECTOMETRYTRANSFORMKIKF_H_
 #define MANTID_MDALGORITHMS_REFLECTOMETRYTRANSFORMKIKF_H_
 
-#include <cmath>
 
 #include "MantidKernel/ClassMacros.h"
 
 #include "MantidAPI/IMDEventWorkspace_fwd.h"
 #include "MantidAPI/MatrixWorkspace_fwd.h"
-
-#include "MantidDataObjects/CalculateReflectometry.h"
+#include "MantidDataObjects/CalculateReflectometryKiKf.h"
 #include "MantidDataObjects/ReflectometryTransform.h"
 
 namespace Mantid {
 namespace MDAlgorithms {
-/**
-class CalculateReflectometryK: Calculation type for converting to ki or kf given
-a theta value (in degrees) and a wavelength
-*/
-class CalculateReflectometryK : public DataObjects::CalculateReflectometry {
-private:
-  double m_sin_theta_i;
-  double m_sin_theta_f;
-
-public:
-  /**
-   Constructor
-   */
-  CalculateReflectometryK() : m_sin_theta_i(0.0), m_sin_theta_f(0.0) {}
-
-  /**
-   Destructor
-   */
-  ~CalculateReflectometryK(){};
-
-  /**
-   Setter for the incident theta value require for the calculation. Internally
-   pre-calculates and caches to cos theta for speed.
-   @param thetaIncident: incident theta value in degrees
-   */
-  void setThetaIncident(double thetaIncident) {
-    m_sin_theta_i = sin(to_radians_factor * thetaIncident);
-  }
-
-  /**
-    Setter for the final theta value require for the calculation. Internally
-    pre-calculates and caches to cos theta for speed.
-    @param thetaFinal: final theta value in degrees
-    */
-  void setThetaFinal(double thetaFinal) {
-    m_sin_theta_f = sin(to_radians_factor * thetaFinal);
-  }
-
-  /**
-   Executes the calculation to determine Ki
-   @param wavelength : wavelength in Angstroms
-   */
-  double calculateDim0(double wavelength) const {
-    double wavenumber = 2 * M_PI / wavelength;
-    return wavenumber * m_sin_theta_i;
-  }
-
-  /**
-   Executes the calculation to determine Kf
-   @param wavelength : wavelength in Angstroms
-   */
-  double calculateDim1(double wavelength) const {
-    double wavenumber = 2 * M_PI / wavelength;
-    return wavenumber * m_sin_theta_f;
-  }
-};
-
 /** ReflectometryTransformKiKf : Type to transform from R vs Wavelength
   workspace to a 2D MDEW with dimensions of Ki and Kf.
