+++ Code/Mantid/Framework/MDAlgorithms/inc/MantidMDAlgorithms/ReflectometryTransformP.h
@@ -1,73 +1,13 @@
 #ifndef MANTID_MDALGORITHMS_REFLECTOMETRYTRANSFORMP_H_
 #define MANTID_MDALGORITHMS_REFLECTOMETRYTRANSFORMP_H_
 
-#include <cmath>
-
-#include "MantidDataObjects/CalculateReflectometry.h"
+#include "MantidDataObjects/CalculateReflectometryP.h"
 #include "MantidDataObjects/ReflectometryTransform.h"
 
 namespace Mantid {
 namespace MDAlgorithms {
 
-/**
-class CalculateReflectometryP: p-type transformation calculator
-*/
-class CalculateReflectometryP : public DataObjects::CalculateReflectometry {
-private:
-  double m_sin_theta_i;
-  double m_sin_theta_f;
 
-public:
-  /**
-   * Constructor
-   */
-  CalculateReflectometryP() : m_sin_theta_i(0.0), m_sin_theta_f(0.0) {}
-
-  /**
-   * Destructor
-   */
-  ~CalculateReflectometryP(){};
-
-  /**
-   Setter for the incident theta value require for the calculation. Internally
-   pre-calculates and caches to cos theta for speed.
-   @param thetaIncident: incident theta value in degrees
-   */
-  void setThetaIncident(double thetaIncident) {
-    m_sin_theta_i = sin(to_radians_factor * thetaIncident);
-  }
-
-  /**
-   Setter for the final theta value require for the calculation. Internally
-   pre-calculates and caches to cos theta for speed.
-   @param thetaFinal: final theta value in degrees
-   */
-  void setThetaFinal(double thetaFinal) {
-    m_sin_theta_f = sin(to_radians_factor * thetaFinal);
-  }
-
-  /**
-   Executes the calculation to determine PSum
-   @param wavelength : wavelength in Angstroms
-   */
-  double calculateDim0(double wavelength) const {
-    double wavenumber = 2 * M_PI / wavelength;
-    double ki = wavenumber * m_sin_theta_i;
-    double kf = wavenumber * m_sin_theta_f;
-    return ki + kf;
-  }
-
-  /**
-   Executes the calculation to determine PDiff
-   @param wavelength : wavelength in Angstroms
-   */
-  double calculateDim1(double wavelength) const {
-    double wavenumber = 2 * M_PI / wavelength;
-    double ki = wavenumber * m_sin_theta_i;
-    double kf = wavenumber * m_sin_theta_f;
-    return ki - kf;
-  }
-};
 
 /** ReflectometryTransformP : Calculates workspace(s) of Pi and Pf based on the
   input workspace and incident theta angle.
