+++ Code/Mantid/Framework/MDAlgorithms/src/ReflectometryTransformQxQz.cpp
@@ -31,7 +31,5 @@ ReflectometryTransformQxQz::ReflectometryTransformQxQz(
   }
   m_calculator->setThetaIncident(incidentTheta);
 }
-
-
 } // namespace Mantid
 } // namespace MDAlgorithms
